Make pre/postrecording int and add logging

+++ machinery/src/models/Config.go
@@ -49,8 +49,8 @@ type Capture struct {
 	USBCamera             USBCamera   `json:"usbcamera"`
 	RaspiCamera           RaspiCamera `json:"raspicamera"`
 	Continuous            string      `json:"continuous,omitempty"`
-	PostRecording         int64       `json:"postrecording"`
-	PreRecording          int64       `json:"prerecording"`
+	PostRecording         int         `json:"postrecording"`
+	PreRecording          int         `json:"prerecording"`
 	MaxLengthRecording    int64       `json:"maxlengthrecording"`
 	TranscodingWebRTC     string      `json:"transcodingwebrtc"`
 	TranscodingResolution int64       `json:"transcodingresolution"`
